+++ src/IfItQuacks.Generator/IsExternalInit.cs
@@ -1,0 +1,6 @@
+namespace System.Runtime.CompilerServices;
+
+// Records need this type, which netstandard2.0 doesn't ship.
+#pragma warning disable S2094
+internal static class IsExternalInit;
+#pragma warning restore S2094